Inline social icons SVG for fewer HTTP requests
This is a possible solution to [#217][0], but I am not sure it is the
"right" change

I used the `asset_data_url` helper to have rails inline a base 64 data
string on assets precompile.

You don't actually *need* to base_64 encode it ([ref][1]), but this was
a quick way to get the desired affect without changing too much of the
asset structure.

[0]:https://github.com/crimethinc/website/issues/217
[1]:https://css-tricks.com/probably-dont-base64-svg/