Add config for capybara and feature specs
This change will add the [capybara gem][0] and set it up so that:
- by default it runs with the `:rack_test` driver (fast)
- if you add `:js` to a spec (e.g. `scenario "foo bar", :js do ...`)
  it will run with `selenium-webdriver` (slow)
- the [`database_cleaner`][1] gem will use transactions for normal specs,
  but truncation for feature specs
- the `launchy` gem is so that when you are debugging a feature spec,
  you can write `save_and_open_screenshot` and a screenshot of the page
  at that point in time will open automatically (note: you must be using
  the `:js` flag)

---

I wanted to write a feature spec for [#347][2], but noticed capybara
wasn't set up yet.

I am not sure it is worth pulling in `database_cleaner` since once
`rspec-rails` catches up with the new minitest API ([wip PR][3]), then
`database_cleaner` will no longer be needed. Maybe it is better to
just wait for that PR to merge (it has 2 ship-its).

[0]:https://github.com/teamcapybara/capybara#drivers
[1]:https://github.com/DatabaseCleaner/database_cleaner#rspec-with-capybara-example
[2]:https://github.com/crimethinc/website/pull/347
[3]:https://github.com/rspec/rspec-rails/pull/1813