[#397] Fix specs
There were 2 things from the 'www' stripping feature that
caused some specs to fail.

First, It looks like the method that a spec was unit
testing (`#strip_www`) was commented out in
8021cf5d2bd56c65653093370950009e6617ec38 and eventually removed
entirely. So this spec is testing something that no longer exists (and
therefore is failing)

I just removed that spec.

Second, capybara/rspec uses a default host of `http://www.example.com`, so
specs using the `redirect_to` matcher were failing (because `www` is now stripped
from requests).

I tried to figure out how to change the default test app host, but nothing
worked and [this rspec issue][0] indicates this is a known, open issue.

Using a full URL in the failing specs gets around the issue, so I did that to
fix the failures.

[0]:https://github.com/rspec/rspec-rails/issues/1275