[TCE] extract font styling to separate file
Why?
------------------------
The last part of creating a turnkey solution for generating TCEs in
additional languages is creating CSS for every new language. One of
the things that make this difficult for someone who isn't familiar
with the codebase is that all of the CSS is together, and you don't
know what and where to look for doing the customizations.

What?
------------------------
 This creates the following structure:
```sh
website/app/assets/stylesheets/to_change_everything
.
├── _common.scss
├── _english.scss
├── _espanol.scss
├── _farsi.scss
├── _한국어.scss
├── common
│   ├── _font_styling.scss
│   ├── _navigation.scss
│   └── _reset.scss
├── farsi
│   └── _font_overrides.scss
└── korean
    └── _font_overrides.scss
```

The top-level files (e.g. `_farsi.scss`) are automatically
generated. So are the
sub-files (e.g. `farsi/_font_overrides.scss`). These are also
automatically wired up and filled in with the same CSS as
`font_styling.scss`.

A person generating a new TCE version then only has to look at 1
file (`farsi/_font_overrides.scss` in this example) and will only
really need to delete any CSS that will be unchanged from the common
styles and adjust some input values in the `scss` functions.

Results
------------------------
Using this convention the custom css needed for the farsi or korean
versions of TCE has been reduced to 50-60 lines of cookie-cutter CSS.

note
------------------------
I do hope to do a similar thing to the layout code (margins, paddings,
etc.) so that there will be a structure like this:

```sh
├── farsi
│   └── _font_overrides.scss
│   └── _layout_overrides.scss

```