Replace jQuery with vanilla JS for 1912 (#1920)
For #1912

As discussed [here](https://github.com/crimethinc/website/commit/ce3cccda5e72e11473ddd83eca4d657a2240d7fd), I have adapted the resize callbacks to be run up to IE8.

Suggestion:
I would suggest to remove the resize events altogether, they shouldn't be of any use on the page. I investigated a bit and it feels like legacy code. The Z-Index applies to titles (there are no blockquotes on the page) that should already be positioned correctly (and it also applies to an element in the header which shouldn't be concerned by this script in my opinion. Removing the functions doesn't seem to affect the layout as far as I've seen.